+++ src/sentineltray/__init__.py
@@ -1,5 +1,5 @@
 """SentinelTray package."""
 
-__version__ = "4.0.0rc1"
-__version_label__ = "4.0.0-rc.1"
+__version__ = "4.0.0rc2"
+__version_label__ = "4.0.0-rc.2"
 __release_date__ = "16-04-2026"